feat: support detach on chain
