+++ crates/replay-chain/src/main.rs
@@ -1,4 +1,5 @@
 mod replay;
+mod setup;
 
 use replay::*;
 
@@ -6,6 +7,7 @@ use anyhow::{anyhow, Context, Result};
 use clap::{App, Arg, SubCommand};
 use gw_config::Config;
 use gw_db::schema::COLUMNS;
+use setup::{setup, SetupArgs};
 use std::{fs, path::Path};
 
 const ARG_CONFIG: &str = "config";
@@ -51,6 +53,37 @@ fn run_cli() -> Result<()> {
                         .required(true),
                 )
                 .display_order(0),
+        )
+        .subcommand(
+            SubCommand::with_name("detach")
+                .about("Detach chain")
+                .arg(
+                    Arg::with_name(ARG_CONFIG)
+                        .short("c")
+                        .takes_value(true)
+                        .required(true)
+                        .default_value("./config.toml")
+                        .help("The config file path"),
+                )
+                .arg(
+                    Arg::with_name("from-db-store")
+                        .long("from-db-store")
+                        .takes_value(true)
+                        .required(true),
+                )
+                .arg(
+                    Arg::with_name("from-db-columns")
+                        .long("from-db-columns")
+                        .takes_value(true)
+                        .required(false),
+                )
+                .arg(
+                    Arg::with_name("to-db-store")
+                        .long("to-db-store")
+                        .takes_value(true)
+                        .required(true),
+                )
+                .display_order(0),
         );
 
     // handle subcommands
@@ -66,13 +99,33 @@ fn run_cli() -> Result<()> {
                 .transpose()?
                 .unwrap_or(COLUMNS);
             let to_db_store = m.value_of("to-db-store").unwrap().into();
-            let args = ReplayArgs {
+            let args = SetupArgs {
                 config,
                 from_db_store,
                 to_db_store,
                 from_db_columns,
             };
-            replay(args).expect("replay");
+            let context = setup(args).expect("setup");
+            replay_chain(context).expect("replay");
+        }
+        ("detach", Some(m)) => {
+            let config_path = m.value_of(ARG_CONFIG).unwrap();
+            let config = read_config(&config_path)?;
+            let from_db_store = m.value_of("from-db-store").unwrap().into();
+            let from_db_columns = m
+                .value_of("from-db-columns")
+                .map(|s| s.parse())
+                .transpose()?
+                .unwrap_or(COLUMNS);
+            let to_db_store = m.value_of("to-db-store").unwrap().into();
+            let args = SetupArgs {
+                config,
+                from_db_store,
+                to_db_store,
+                from_db_columns,
+            };
+            let context = setup(args).expect("setup");
+            detach_chain(context).expect("detach");
         }
         _ => {
             return Err(anyhow!("unknown command"));
